requirements.txt
Django==3.0
Pillow==7.2.0
psycopg2-binary==2.7.7
social-auth-app-django==4.0.0
django-imagekit==4.0.2
